+++ frontend/apps/web/app/(mgmt)/[account]/transformers/EditTransformerOptions.tsx
@@ -38,7 +38,7 @@ import { useForm } from 'react-hook-form';
 import {
   EditJobMappingTransformerConfigFormContext,
   EditJobMappingTransformerConfigFormValues,
-} from '../new/transformer/schema';
+} from '../new/transformer/transform-form-validations';
 import TransformerForm from '../new/transformer/TransformerForms/TransformerForm';
 
 interface Props {
